refactor: Move callback excep to callback module
