+++ dl_utils/learner.py
@@ -1,27 +1,16 @@
 import torch
 
-from .callbacks import ProgressCallback, Recorder, TrainEvalCallback
+from .callbacks import (
+    CancelBatchException,
+    CancelEpochException,
+    CancelFitException,
+    CancelTrainException,
+    CancelValidException,
+    ProgressCallback,
+    Recorder,
+    TrainEvalCallback,
+)
 from .utils import listify, setify
-
-
-class CancelFitException(Exception):
-    """Stop training and exit"""
-
-
-class CancelEpochException(Exception):
-    """Stop current epoch and move to next epoch."""
-
-
-class CancelTrainException(Exception):
-    """Stop training current batch and move to validation."""
-
-
-class CancelValidException(Exception):
-    """Stop validation phase and move to next epoch"""
-
-
-class CancelBatchException(Exception):
-    """Stop current batch and move to next batch."""
 
 
 def params_getter(model):
